Removing WeblogEntry interface, fixing 1.4.2 compile problems
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/roller/branches/roller_3.0@412521 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/roller/business/hibernate/HibernateWeblogManagerImpl.java
@@ -1068,7 +1068,7 @@ public class HibernateWeblogManagerImpl implements WeblogManager {
                     (String)row[2], 
                     (String)row[3], 
                     "statCount.weblogEntryCommentCountType", 
-                    new Long((Integer)row[0]).longValue()));
+                    new Long(((Integer)row[0]).intValue()).longValue()));
             }
             return results;
         } catch (Throwable pe) {
